extractOrderNumber export change, advocateName fix in multiple adv case (#2246)
