+++ utilities/dristi-pdf/src/applicationHandlers/applicationRescheduleRequest.js
@@ -2,15 +2,12 @@ const cheerio = require("cheerio");
 const config = require("../config");
 const {
   search_case,
-  search_mdms,
-  search_hrms,
   search_sunbirdrc_credential_service,
   search_application,
   create_pdf,
   search_advocate,
 } = require("../api");
 const { renderError } = require("../utils/renderError");
-const { getAdvocates } = require("./getAdvocates");
 const { cleanName } = require("./cleanName");
 
 function getOrdinalSuffix(day) {
@@ -86,7 +83,6 @@ async function applicationRescheduleRequest(req, res, qrCode) {
     if (!courtCase) {
       return renderError(res, "Court case not found", 404);
     }
-    const allAdvocates = getAdvocates(courtCase);
 
     // Search for HRMS details
     // const resHrms = await handleApiCall(
@@ -144,6 +140,7 @@ async function applicationRescheduleRequest(req, res, qrCode) {
     }
 
     let barRegistrationNumber = "";
+    let advocateName = "";
     const advocateIndividualId =
       application?.applicationDetails?.advocateIndividualId;
     if (advocateIndividualId) {
@@ -156,14 +153,11 @@ async function applicationRescheduleRequest(req, res, qrCode) {
         (item) => item.isActive === true
       );
       barRegistrationNumber = advocateDetails?.barRegistrationNumber || "";
+      advocateName =
+        cleanName(advocateDetails?.additionalDetails?.username) || "";
     }
 
     const onBehalfOfuuid = application?.onBehalfOf?.[0];
-    const advocate = allAdvocates[onBehalfOfuuid]?.[0]?.additionalDetails
-      ?.advocateName
-      ? allAdvocates[onBehalfOfuuid]?.[0]
-      : {};
-    const advocateName = cleanName(advocate?.additionalDetails?.advocateName || "");
     const partyName = application?.additionalDetails?.onBehalOfName || "";
     const onBehalfOfLitigent = courtCase?.litigants?.find(
       (item) => item.additionalDetails.uuid === onBehalfOfuuid
