+++ utilities/dristi-pdf/src/applicationHandlers/caseSettlementApplication.js
@@ -2,16 +2,12 @@ const cheerio = require("cheerio");
 const config = require("../config");
 const {
   search_case,
-  search_mdms,
-  search_hrms,
   search_sunbirdrc_credential_service,
   search_application,
   create_pdf,
   search_advocate,
 } = require("../api");
 const { renderError } = require("../utils/renderError");
-const { getAdvocates } = require("./getAdvocates");
-const { formatDate } = require("./formatDate");
 const { cleanName } = require("./cleanName");
 
 function getOrdinalSuffix(day) {
@@ -84,7 +80,6 @@ async function caseSettlementApplication(req, res, qrCode) {
     if (!courtCase) {
       return renderError(res, "Court case not found", 404);
     }
-    const allAdvocates = getAdvocates(courtCase);
 
     // Search for HRMS details
     // const resHrms = await handleApiCall(
@@ -141,6 +136,7 @@ async function caseSettlementApplication(req, res, qrCode) {
     }
 
     let barRegistrationNumber = "";
+    let advocateName = "";
     const advocateIndividualId =
       application?.applicationDetails?.advocateIndividualId;
     if (advocateIndividualId) {
@@ -153,14 +149,11 @@ async function caseSettlementApplication(req, res, qrCode) {
         (item) => item.isActive === true
       );
       barRegistrationNumber = advocateDetails?.barRegistrationNumber || "";
+      advocateName =
+        cleanName(advocateDetails?.additionalDetails?.username) || "";
     }
 
     const onBehalfOfuuid = application?.onBehalfOf?.[0];
-    const advocate = allAdvocates?.[onBehalfOfuuid]?.[0]?.additionalDetails
-      ?.advocateName
-      ? allAdvocates[onBehalfOfuuid]?.[0]
-      : {};
-    const advocateName = cleanName(advocate?.additionalDetails?.advocateName || "");
     const partyName = application?.additionalDetails?.onBehalOfName || "";
     const onBehalfOfLitigent = courtCase?.litigants?.find(
       (item) => item.additionalDetails.uuid === onBehalfOfuuid
@@ -221,22 +214,22 @@ async function caseSettlementApplication(req, res, qrCode) {
       "November",
       "December",
     ];
-    const applicationNameMap = {
-      BAIL_BOND: "Bail Application - Personal Bail Bond",
-      SURETY: "Bail Application - In Person Surety",
-      CHECKOUT_REQUEST: "Checkout Application",
-      SETTLEMENT: "Case Settlement Application",
-      TRANSFER: "Case Transfer Application",
-      WITHDRAWAL: "Case Withdrawal",
-      PRODUCTION_DOCUMENTS:
-        "Application for production of documents or evidence",
-      EXTENSION_SUBMISSION_DEADLINE: "Application for Extension of Submission",
-      "": "General Application",
-      undefined: "General Application",
-    };
+    // const applicationNameMap = {
+    //   BAIL_BOND: "Bail Application - Personal Bail Bond",
+    //   SURETY: "Bail Application - In Person Surety",
+    //   CHECKOUT_REQUEST: "Checkout Application",
+    //   SETTLEMENT: "Case Settlement Application",
+    //   TRANSFER: "Case Transfer Application",
+    //   WITHDRAWAL: "Case Withdrawal",
+    //   PRODUCTION_DOCUMENTS:
+    //     "Application for production of documents or evidence",
+    //   EXTENSION_SUBMISSION_DEADLINE: "Application for Extension of Submission",
+    //   "": "General Application",
+    //   undefined: "General Application",
+    // };
 
     const currentDate = new Date();
-    const formattedToday = formatDate(currentDate, "DD-MM-YYYY");
+    // const formattedToday = formatDate(currentDate, "DD-MM-YYYY");
     const day = currentDate.getDate();
     const month = months[currentDate.getMonth()];
     const year = currentDate.getFullYear();
