+++ utilities/dristi-pdf/src/utils/orderUtils.js
@@ -312,18 +312,10 @@ async function handleCompositePDF(req, res, qrCode, order) {
   }
 }
 
-const extractOrderNumber = (orderItemId) => {
-  if (!orderItemId || typeof orderItemId !== "string") return orderItemId || "";
-  return orderItemId?.includes("_")
-    ? orderItemId?.split("_")?.pop()
-    : orderItemId;
-};
-
 module.exports = {
   OrderPreviewOrderTypeMap,
   orderPDFMap,
   applicationStatusType,
   processOrder,
   handleCompositePDF,
-  extractOrderNumber,
 };
